+++ src/entities/announcement/api/use-recent-viewed.queries.ts
@@ -1,6 +1,6 @@
 import { useInfiniteQuery } from "@tanstack/react-query";
 import { Api } from "@/src/shared/api/api";
-import { RecentViewedResponseSchema } from "@/src/entities/announcement/model/schemas";
+import { RecentViewedResponseSchema } from "@/src/entities/announcement/model/announcement.schemas";
 
 export function useRecentViewedAnnouncements(limit = 20) {
   return useInfiniteQuery({
