+++ src/features/filter-announcements/model/use-filter-store.ts
@@ -1,41 +1,5 @@
 import { create } from "zustand";
-import type { AnnouncementFilterParams } from "@/src/entities/announcement/model/types";
-
-type SortType = "DEADLINE" | "LATEST" | "RELEVANCE";
-type FilterTab = "region" | "publisher" | "housingType";
-type StatusType = "OPEN" | "CLOSED";
-
-interface FilterState {
-  isFilterOpen: boolean;
-  activeTab: FilterTab;
-  statusTab: StatusType;
-  isPersonalized: boolean;
-
-  tempFilters: {
-    regionName: string | undefined;
-    publisher?: string;
-    housingType?: string;
-    keyword?: string;
-  };
-  appliedFilters: AnnouncementFilterParams;
-
-  toggleFilter: (tab: FilterTab) => void;
-  closeFilter: () => void;
-  setActiveTab: (tab: FilterTab) => void;
-  setStatusTab: (status: StatusType) => void;
-  setIsPersonalized: (enabled: boolean) => void;
-  setTempFilter: (
-    key: keyof FilterState["tempFilters"],
-    value: string | undefined,
-  ) => void;
-  setFilter: <K extends keyof AnnouncementFilterParams>(
-    key: K,
-    value: AnnouncementFilterParams[K],
-  ) => void;
-  setSort: (sort: SortType) => void;
-  applyFilters: () => void;
-  resetFilters: () => void;
-}
+import type { FilterState } from "@/src/features/filter-announcements/model/filter.types";
 
 const initialFilterState = {
   regionName: undefined,
@@ -61,11 +25,8 @@ export const useFilterStore = create<FilterState>((set) => ({
     }),
 
   closeFilter: () => set({ isFilterOpen: false }),
-
   setActiveTab: (activeTab) => set({ activeTab }),
-
   setStatusTab: (statusTab) => set({ statusTab }),
-
   setIsPersonalized: (isPersonalized) => set({ isPersonalized }),
 
   setSort: (sort) =>
