+++ src/widgets/announcement-list/ui/announcement-list.tsx
@@ -2,7 +2,7 @@
 
 import Link from "next/link";
 import { useAnnouncements } from "@/src/entities/announcement/api/use-announcements.queries";
-import type { Announcement } from "@/src/entities/announcement/model/types";
+import type { Announcement } from "@/src/entities/announcement/model/announcement.types";
 import { useFilterStore } from "@/src/features/filter-announcements/model/use-filter-store";
 import { AnnouncementCard } from "@/src/widgets/announcement-card/ui/announcement-card";
 import { useEffect } from "react";
